Fix bug with nested rules checking
- There was a bug where a search for a nested rule only checks the
children of the first instance of the top node that it finds in the
ast tree. Fixed this by implementing findNodeWithType method using a
generator, so that it iterates over all instances of a node type in
the ast tree. It remains to be checked if this will cayse significant
performance regression with deep nested rule checks.